Attributes Stuff.
Attributes are now loaded by all entities. Although it might be better if I stored it as an interface for certain items, however that would add BS, and complexity so I guess this is best. Move away from using strings to store value of attributes, and now they are stored as numbers.