Implement testing of requirements.
Refactor how the valid method will work. Now any thing deriving from entity base can be passed in, and it won't matter if it is an inventory till the Item gets ahold of it and does the check.
Expand test cases.
Implement the go through and test part of the requirement check system.
Start work on printing out internals of the Requirement engine.